+++ .gitignore
@@ -1,4 +1,5 @@
 /.env
+.vscode/
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
